+++ base/src/AzAudio/AzAudio.c
@@ -49,7 +49,7 @@ int azaInit() {
 			}
 		}
 	}
-	AZA_LOG_INFO("AzAudio Version: " AZA_VERSION_FORMAT_STR "\n", AZA_VERSION_ARGS);
+	AZA_LOG_INFO("AzAudio Version: %s\n", azaVersionString);
 
 	int err;
 	// A resolution of 128 is 2^7, which gives the LUT a signal-to-noise ratio of 12+12*7 = 96dB